Extract BraveBatchPassageEmbedder to its own .h/.cc
Pure file split. The class declaration and definitions move from
brave_passage_embeddings_service.{h,cc} into brave_batch_passage_embedder.{h,cc}
unchanged. Service.h forward-declares BraveBatchPassageEmbedder; the
include lives in service.cc only. The history_embeddings GN target
gets a public/sources split so brave_batch_passage_embedder.h is
internal to the target.

Adds brave_batch_passage_embedder_unittest.cc covering the
BatchEmbedder's surface (batch ordering, sequential serialization,
empty/out-of-range renderer responses, receiver/renderer disconnect
firing on_disconnect).
